Test: Split tests to folders
