+++ workspace/test/schema/__init__.py
@@ -1,0 +1,1 @@
+"""Workspace schema tests."""
